feat-[167165603]: Create Reuseable Components
- Update Comment Card
- Update CreateComment Card
- Update Search Results Card

- Update Suggested Articles Card

[Delivers #167165603]
