+++ src/components/common/displayStar.jsx
@@ -1,50 +1,39 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
+import FontAwesome from './FontAwesome';
 
 export default function DisplayStar({ averageRating }) {
   const starRating = Math.floor(parseInt(averageRating, 10));
   if (starRating === 4) {
     return (
       <Fragment>
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star-o" />
+        <FontAwesome />
+        <FontAwesome />
+        <FontAwesome />
+        <FontAwesome />
       </Fragment>
     );
   }
   if (starRating === 3) {
     return (
       <Fragment>
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
+        <FontAwesome />
+        <FontAwesome />
+        <FontAwesome />
       </Fragment>
     );
   }
   if (starRating === 2) {
     return (
       <Fragment>
-        <i className="fa fa-star" />
-        <i className="fa fa-star" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
+        <FontAwesome />
+        <FontAwesome />
       </Fragment>
     );
   }
   if (starRating === 1) {
     return (
-      <Fragment>
-        <i className="fa fa-star" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
-        <i className="fa fa-star-o" />
-      </Fragment>
+      <FontAwesome />
     );
   }
   if (starRating === 0) {
@@ -52,11 +41,11 @@ export default function DisplayStar({ averageRating }) {
   }
   return (
     <Fragment>
-      <i className="fa fa-star" />
-      <i className="fa fa-star" />
-      <i className="fa fa-star" />
-      <i className="fa fa-star" />
-      <i className="fa fa-star" />
+      <FontAwesome />
+      <FontAwesome />
+      <FontAwesome />
+      <FontAwesome />
+      <FontAwesome />
     </Fragment>
   );
 }
