+++ src/components/common/readTime.jsx
@@ -3,11 +3,9 @@ import PropTypes from "prop-types";
 
 export default function ReadTime({ time }) {
   return (
-    <>
-      <p className="text-black text-gray-600 text-sm bold font-sans w-full">
-        {time}
-      </p>
-    </>
+    <p className="text-black text-gray-600 text-sm bold font-sans w-full">
+      {time}
+    </p>
   );
 }
