+++ src/index.html
@@ -8,7 +8,6 @@
   <meta name="keywords" content="React.js, JSX">
   <meta name="author" content="Andela Simulations Programme">
   <link href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
   <title>Author's Haven</title>
 </head>
 <body>
